feat: restrict guest review deletion, reload on user change/logout (#10)
Suggestions by @adriahso in PR #15
The bug where a logged out user (Guest) could not delete a review, is now an intended feature.
Changing username or logging out will refresh the page to show the new user's available actions. This is a temporary solution until we have global state management.