build(ci): specify AVX runner for MongoDB Memory Server support (#69)
Force action to run on AVX compatible hardware to ensure compatibility with MongoDB Memory Server, which relies on MongoDB 5.0+ (which requires AVX support).
The already existing runners are hosted on a machine without AVX support. A new runner was set up on our VM that includes AVX capabilities and labeled accordingly.